v22.5.2 - Month/year wheel picker, sync reliability, reminder improvements
Features:
- Inline month/year wheel picker in date picker (fixes #52)
- Circular wheel pickers for month and year navigation
- Done button on year overlay for explicit confirmation
- Show absolute event time in reminder notifications

Fixes:
- Skip FK constraint errors during sync instead of aborting (fixes #55)
- Month wheel picker initial render selection
- Disable NullSafeMutableLiveData lint crash on lifecycle 2.8.7
- Prevent int overflow in time calculations
- Correct F-Droid changelog version code

Docs:
- Add new F-Droid screenshots (wheel picker, view options, color picker, account management)
- Update README screenshot table
- Add v22.5.2 F-Droid changelog

+++ app/src/main/kotlin/org/onekash/kashcal/data/db/entity/IcsSubscription.kt
@@ -142,7 +142,7 @@ data class IcsSubscription(
     fun isDueForSync(): Boolean {
         if (!enabled) return false
         val now = System.currentTimeMillis()
-        val intervalMs = syncIntervalHours * 60 * 60 * 1000L
+        val intervalMs = syncIntervalHours.toLong() * 60 * 60 * 1000
         return now - lastSync >= intervalMs
     }
